refactor: WarningLevelCalculator의 메서드를 WarningLevel로 위임
- WarningLevel을 반환하는 메서드를 WarningLevel에서 관리하도록 변경
- Predicate를 Enum의 필드로 두어 WarningLevel 반환 검증에 사용
