+++ src/test/java/attendance/domain/AttendanceCheckerTest.java
@@ -1,11 +1,9 @@
 package attendance.domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.format.TextStyle;
 import java.util.Locale;
 import org.junit.jupiter.api.DisplayName;
